Split header and source
